+++ crates/ruma-common/CHANGELOG.md
@@ -17,6 +17,7 @@ Improvements:
   * Move `Relation::Thread` and associated types and methods out of `unstable-msc3440`
   * Add parameter to `RoomMessageEventContent::make_reply_to` to be thread-aware
   * Add `RoomMessageEventContent::make_for_reply`
+* Stabilize support for event replacements (edits)
 
 # 0.10.3